pin localisation et zoom sur la localisation

+++ Final/app/src/main/java/ca/ulaval/ima/mp/model/DataClass.kt
@@ -72,8 +72,8 @@ data class RestaurantLight(
 )
 
 data class Location(
-        @SerializedName("latitude") val latitude:Float,
-        @SerializedName("longitude") val longitude:Float
+        @SerializedName("latitude") val latitude:Double,
+        @SerializedName("longitude") val longitude:Double
 )
 
 data class OpeningHour(
